Rename the test logic
