+++ frontend/src/scenes/billing/v2/test/BillingLocked.tsx
@@ -1,5 +1,5 @@
 import { useValues } from 'kea'
-import { billingLogic } from './billingLogic'
+import { billingLogic } from './billingTestLogic'
 import { Billing } from '../../Billing'
 import { AlertMessage } from 'lib/components/AlertMessage'
 import './BillingLocked.scss'
